+++ examples/portfolio/native.jl
@@ -1,10 +1,7 @@
 #=
 Copyright 2019, Chris Coey, Lea Kapelevich and contributors
 
-TODO
-- add description
-- add more risk constraints covering more cones
-- enable random generation
+maximize expected returns subject to risk constraints
 =#
 
 using LinearAlgebra
@@ -25,9 +22,13 @@ function portfolio(
     returns::Vector{Float64} = Float64[],
     sigma_half::AbstractMatrix{Float64} = Matrix{Float64}(undef, 0, 0),
     gamma::Float64 = -1.0,
-    risk_measure::Symbol = :entropic,
+    risk_measures::Vector{Symbol} = [:entropic],
     use_l1ball::Bool = true,
+    use_linfball::Bool = true,
     )
+    if :entropic in risk_measures && length(risk_measures) > 1
+        error("we currently assume entropic ball doesn't intersect with another ball")
+    end
     if isempty(returns)
         returns = rand(num_stocks)
     end
@@ -38,52 +39,91 @@ function portfolio(
     end
     if gamma < 0
         x = randn(num_stocks)
-        gamma = sum(abs, sigma_half * x)
+        x ./= norm(x)
+        gamma = sum(abs, sigma_half * x) / sqrt(num_stocks)
     end
 
     c = returns
+    # investments add to one
     A = ones(1, num_stocks)
     b = [1.0]
-    G1 = -Matrix{Float64}(I, num_stocks, num_stocks)
-    h1 = zeros(num_stocks)
-    cone_idxs = [1:num_stocks]
+    # nonnegativity
+    G = -Matrix{Float64}(I, num_stocks, num_stocks)
+    h = zeros(num_stocks)
+    cone_idxs = UnitRange{Int}[1:num_stocks]
     cones = CO.Cone[CO.Nonnegative{Float64}(num_stocks)]
+    cone_offset = num_stocks
 
-    if risk_measure == :quadratic || (risk_measure == :l1 && use_l1ball)
-        G2 = vcat(zeros(1, num_stocks), -sigma_half)
-        h2 = [gamma; zeros(num_stocks)]
-        cone_idxs = vcat(cone_idxs, [(num_stocks + 1):(2 * num_stocks + 1)])
-        G = vcat(G1, G2)
-        h = vcat(h1, h2)
-        if risk_measure == :quadratic
-            cones = vcat(cones, CO.EpiNormEucl{Float64}(num_stocks + 1))
-        else
-            cones = vcat(cones, CO.EpiNormInf{Float64}(num_stocks + 1, true))
-        end
+    function add_single_ball(cone, gamma)
+        G_risk = vcat(zeros(1, num_stocks), -sigma_half)
+        h_risk = [gamma; zeros(num_stocks)]
+        G = vcat(G, G_risk)
+        h = vcat(h, h_risk)
+        push!(cones, cone)
+        push!(cone_idxs, (cone_offset + 1):(cone_offset + num_stocks + 1))
+        cone_offset += num_stocks + 1
+    end
 
-    elseif risk_measure == :l1 && !use_l1ball
+    if :quadratic in risk_measures
+        add_single_ball(CO.EpiNormEucl{Float64}(num_stocks + 1), gamma)
+    end
+    if :l1 in risk_measures && use_l1ball
+        add_single_ball(CO.EpiNormInf{Float64}(num_stocks + 1, true), gamma * sqrt(num_stocks))
+    end
+    if :linf in risk_measures && use_linfball
+        add_single_ball(CO.EpiNormInf{Float64}(num_stocks + 1), gamma)
+    end
+
+    if :l1 in risk_measures && !use_l1ball
         c = vcat(c, zeros(2 * num_stocks))
         id = Matrix{Float64}(I, num_stocks, num_stocks)
         id2 = Matrix{Float64}(I, 2 * num_stocks, 2 * num_stocks)
         A_slacks = [sigma_half -id id]
         A_l1 = [zeros(1, num_stocks) ones(1, 2 * num_stocks)]
-        A = [A zeros(1, 2 * num_stocks); A_slacks; A_l1]
-        b = vcat(b, zeros(num_stocks), gamma)
+        A = [
+            A zeros(1, 2 * num_stocks)
+            A_slacks
+            A_l1
+            ]
+        b = vcat(b, zeros(num_stocks), gamma * sqrt(num_stocks))
         G = [
-            G1 zeros(num_stocks, 2 * num_stocks)
+            G zeros(size(G, 1), 2 * num_stocks)
             zeros(2 * num_stocks, num_stocks) -id2
             ]
-        h = vcat(h1, zeros(2 * num_stocks))
-        cones = vcat(cones, CO.Nonnegative{Float64}(2 * num_stocks))
-        cone_idxs = vcat(cone_idxs, [(num_stocks + 1):(3 * num_stocks)])
+        h = vcat(h, zeros(2 * num_stocks))
+        push!(cones, CO.Nonnegative{Float64}(2 * num_stocks))
+        push!(cone_idxs, (cone_offset + 1):(cone_offset + 2 * num_stocks))
+        cone_offset += 2 * num_stocks
+    end
 
-    elseif risk_measure == :entropic
+    if :linf in risk_measures && !use_linfball
+        c = vcat(c, zeros(2 * num_stocks))
+        id = Matrix{Float64}(I, num_stocks, num_stocks)
+        fill_cols = size(A, 2) - num_stocks
+        fill_rows = size(A, 1)
+        A = [
+            A zeros(fill_rows, 2 * num_stocks)
+            sigma_half zeros(num_stocks, fill_cols) id zeros(num_stocks, num_stocks)
+            -sigma_half zeros(num_stocks, fill_cols) zeros(num_stocks, num_stocks) id
+            ]
+        b = vcat(b, gamma * ones(2 * num_stocks))
+        G = [
+            G zeros(size(G, 1), 2 * num_stocks)
+            zeros(2 * num_stocks, size(G, 2)) -Matrix{Float64}(I, 2 * num_stocks, 2 * num_stocks)
+            ]
+        h = vcat(h, zeros(2 * num_stocks))
+        push!(cones, CO.Nonnegative{Float64}(2 * num_stocks))
+        push!(cone_idxs, (cone_offset + 1):(cone_offset + 2 * num_stocks))
+        cone_offset += 2 * num_stocks
+    end
+
+    if :entropic in risk_measures
         # sigma_half = abs.(sigma_half) TODO will this always be feasible?
         c = vcat(c, zeros(2 * num_stocks))
-        A = [A zeros(1, 2 * num_stocks); zeros(1, num_stocks) ones(1, 2 * num_stocks)]
+        A = [A zeros(size(A, 1), 2 * num_stocks); zeros(1, num_stocks) ones(1, 2 * num_stocks)]
         b = vcat(b, gamma^2)
-        G2pos = zeros(3 * num_stocks, 3 * num_stocks)
-        G2neg = zeros(3 * num_stocks, 3 * num_stocks)
+        G2pos = zeros(3 * num_stocks, 2 * num_stocks + size(G, 2))
+        G2neg = zeros(3 * num_stocks, 2 * num_stocks + size(G, 2))
         h2 = zeros(3 * num_stocks)
 
         offset = 1
@@ -96,12 +136,11 @@ function portfolio(
             G2neg[offset + 1, 1:num_stocks] = sigma_half[i, :]
             offset += 3
         end
-        G = [G1 zeros(num_stocks, 2 * num_stocks); G2pos; G2neg]
-        h = vcat(h1, h2, h2)
-        cone_idxs = vcat(cone_idxs, [(3 * (i - 1) + num_stocks + 1):(3 * i + num_stocks) for i in 1:(2 * num_stocks)])
+        G = [G zeros(size(G, 1), 2 * num_stocks); G2pos; G2neg]
+        h = vcat(h, h2, h2)
+        cone_idxs = vcat(cone_idxs, [(3 * (i - 1) + cone_offset + 1):(3 * i + cone_offset) for i in 1:(2 * num_stocks)])
         cones = vcat(cones, [CO.HypoPerLog{Float64}() for _ in 1:(2 * num_stocks)])
-    else
-        error("unknown risk measure: $risk_measure")
+        cone_offset += 6 * num_stocks
     end
 
     return (c = c, A = A, b = b, G = G, h = h, cones = cones, cone_idxs = cone_idxs, sigma_half = sigma_half, gamma = gamma)
@@ -114,12 +153,12 @@ function test_portfolio(instance::Function; options, rseed::Int = 1)
     solver = SO.HSDSolver{Float64}(model; options...)
     SO.solve(solver)
     r = SO.get_certificates(solver, model, test = true, atol = 1e-4, rtol = 1e-4)
+    @show d.sigma_half * r.x[1:size(d.sigma_half, 1)]
     # @test r.status == :Optimal
-    # @assert sum(abs, d.sigma_half * r.x) <= d.gamma
     return
 end
 
-portfolio_ex(risk_measure::Symbol; kwargs...) = portfolio(
+portfolio_ex(risk_measures::Vector{Symbol}; kwargs...) = portfolio(
     3,
     returns = -[0.0254,	0.0190,	0.0045],
     sigma_half = cholesky([
@@ -128,25 +167,35 @@ portfolio_ex(risk_measure::Symbol; kwargs...) = portfolio(
         0.0001	0.0002	0.0019
         ]).U,
     gamma = 0.033,
-    risk_measure = risk_measure;
+    risk_measures = risk_measures;
     kwargs...,
     )
 
-portfolio1() = portfolio_ex(:quadratic)
-portfolio2() = portfolio_ex(:entropic)
-portfolio3() = portfolio(4, risk_measure = :l1, use_l1ball = true)
-portfolio4() = portfolio(4, risk_measure = :l1, use_l1ball = false)
+portfolio1() = portfolio_ex([:quadratic])
+portfolio2() = portfolio_ex([:entropic])
+portfolio3() = portfolio(4, risk_measures = [:l1], use_l1ball = true)
+portfolio4() = portfolio(4, risk_measures = [:l1], use_l1ball = false)
+portfolio5() = portfolio(4, risk_measures = [:linf], use_linfball = true)
+portfolio6() = portfolio(4, risk_measures = [:linf], use_linfball = false)
+portfolio7() = portfolio(4, risk_measures = [:linf, :l1], use_linfball = true, use_l1ball = true)
+portfolio8() = portfolio(4, risk_measures = [:linf, :l1], use_linfball = false, use_l1ball = false)
 
 test_portfolio_all(; options...) = test_portfolio.([
     portfolio1,
     portfolio2,
+    portfolio3,
+    portfolio4,
+    portfolio5,
+    portfolio6,
+    portfolio7,
+    portfolio8,
     ], options = options)
 
 test_portfolio(; options...) = test_portfolio.([
     portfolio1,
     portfolio2,
-    portfolio3,
-    portfolio4,
+    portfolio7,
+    portfolio8,
     ], options = options)
 
 
@@ -164,7 +213,7 @@ test_portfolio(; options...) = test_portfolio.([
 #     for use_l1ball in tf
 #
 #         Random.seed!(seed)
-#         d = portfolio(n, risk_measure = :l1, use_l1ball = use_l1ball)
+#         d = portfolio(n, risk_measures = [:l1], use_l1ball = use_l1ball)
 #         model = MO.PreprocessedLinearModel{T}(d.c, d.A, d.b, d.G, d.h, d.cones, d.cone_idxs)
 #         solver = SO.HSDSolver{T}(model, tol_abs_opt = 1e-5, tol_rel_opt = 1e-5, time_limit = 600)
 #         t = @timed SO.solve(solver)
